+++ Sources/OpenAI/Public/Types/ChatVisionQuery.swift
@@ -84,14 +84,14 @@ public enum ChatVisionMessage: Codable, Equatable {
 public struct ChatVisionQuery: Equatable, Codable, Streamable {
     public let model: String
     public let messages: [ChatVisionMessage]
-    public let temperature: Double?
-    public let topP: Double?
+    public let temperature: Float?
+    public let topP: Float?
     public let n: Int?
     public let seed: Int?
     public let stop: [String]?
     public let maxTokens: Int?
-    public let presencePenalty: Double?
-    public let frequencyPenalty: Double?
+    public let presencePenalty: Float?
+    public let frequencyPenalty: Float?
     public let logitBias: [String:Int]?
     public let user: String?
     
@@ -113,9 +113,9 @@ public struct ChatVisionQuery: Equatable, Codable, Streamable {
         case user
     }
     
-    public init(model: String, messages: [ChatVisionMessage], temperature: Double? = nil, topP: Double? = nil, n: Int? = nil,
-                seed: Int? = nil, stop: [String]? = nil, maxTokens: Int? = nil, presencePenalty: Double? = nil,
-                frequencyPenalty: Double? = nil, logitBias: [String : Int]? = nil, user: String? = nil, stream: Bool = false) {
+    public init(model: String, messages: [ChatVisionMessage], temperature: Float? = nil, topP: Float? = nil, n: Int? = nil,
+                seed: Int? = nil, stop: [String]? = nil, maxTokens: Int? = nil, presencePenalty: Float? = nil,
+                frequencyPenalty: Float? = nil, logitBias: [String : Int]? = nil, user: String? = nil, stream: Bool = false) {
         self.model = model
         self.messages = messages
         self.temperature = temperature
